# docs/plan/syncnode/openapi.yaml
openapi: 3.0.3
info:
  title: CubeFS Syncnode — Complete Admin API
  version: "2.0.0"
  description: |
    Complete HTTP admin surface for the CubeFS sync subsystem.

    **Two servers, two surfaces:**

    | Server | Default Port | Surface |
    |--------|-------------|---------|
    | Master | 17010 | All sync management APIs: node registration, rule CRUD, task observability, direct dispatch |
    | Syncnode | 17911 | Local diagnostics only: version, stat, reload |

    **Authentication:**
    All master sync endpoints require `Authorization: Bearer <syncAdminToken>` (or
    `X-Sync-Token: <token>`). The token is configured as `syncAdminToken` in
    `master.json`. An empty value disables the check (dev/test mode).

    Syncnode local endpoints share the same token scheme but are only
    reachable on the node itself.

    **Response envelope (master + syncnode):**
    ```json
    { "code": 0, "msg": "OK", "data": <payload> }
    ```
    Non-zero `code` signals an error; `data` is null. HTTP status is
    always 200 — check `code` in the body.

    **Common error codes:**
    | code | meaning |
    |------|---------|
    | 0 | success |
    | 7 | ErrCodeParamError — missing or malformed param / body |
    | 10 | ErrCodePersistenceByRaft — raft submit failed; safe to retry |
    | 11 | ErrCodeInternalError — not found, conflict, or other server error |

  contact:
    name: CubeFS Sync Team
servers:
  - url: "http://{host}:{masterPort}"
...
    variables:
      host:
...
security:
  - BearerAuth: []
  - SyncTokenAuth: []

# ────────────────────────────────────────────────────────────────────────────
# Master: SyncNode registration + listing
# ────────────────────────────────────────────────────────────────────────────
paths:
...
  /syncNode/decommission:
    post:
      tags: [node-mgmt]
      summary: 下线 syncnode
      description: |
        State transitions:
        - `force=false` (default): mark the node `draining` and return.
          Cluster cron removes it when active task count drops to zero.
          Operator monitors via `GET /syncNode/tasks`.
        - `force=true`: mark draining + immediately cancel/failover every
          active task on the node, then raft-delete the node record.

        Node state machine:
        ```
        Active ──/decommission?force=false──→ Draining ──(tasks=0)──→ Removed
        Active ──/decommission?force=true ──→ Draining ──(failover)──→ Removed
        ```
      operationId: decommissionSyncNode
      parameters:
        - name: addr
          in: query
          required: true
          schema:
            type: string
            example: "10.89.6.10:17911"
        - name: force
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Response"
                  - properties:
                      data:
                        $ref: "#/components/schemas/NodeLifecycleResponse"
  /syncNode/drain:
    post:
      tags: [node-mgmt]
      summary: 排空 syncnode 任务（保留注册）
      description: |
        Cancels and failovers every active task on the node but keeps the
        node registered. Use when preparing for maintenance; follow up with
        `/syncNode/restore` to return the node to active scheduling.

        State machine:
        ```
        Active ──/drain──→ Draining ──/restore──→ Active
        ```
      operationId: drainSyncNode
      parameters:
        - name: addr
          in: query
          required: true
          schema:
            type: string
            example: "10.89.6.10:17911"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Response"
                  - properties:
                      data:
                        $ref: "#/components/schemas/NodeLifecycleResponse"
  /syncNode/restore:
    post:
      tags: [node-mgmt]
      summary: 恢复已排空节点为 active
      description: |
        Flips the node state from `draining` back to `active`. Idempotent
        if the node is already `active`. No effect on running tasks —
        cancellations were dispatched by `/drain`; new tasks will start
        being scheduled to the node again immediately.
      operationId: restoreSyncNode
      parameters:
        - name: addr
          in: query
          required: true
          schema:
            type: string
            example: "10.89.6.10:17911"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Response"
                  - properties:
                      data:
                        type: object
                        properties:
                          addr:
                            type: string
                          state:
                            type: string
                            enum: [active, draining]
  /syncNode/tasks:
    get:
      tags: [node-mgmt]
      summary: 查询指定节点上的任务列表
      description: |
        Returns the task ledger entries whose `owner` matches the given addr.
        Useful for monitoring drain progress — poll until the list is empty.
      operationId: listSyncNodeTasks
      parameters:
        - name: addr
          in: query
          required: true
          schema:
            type: string
            example: "10.89.6.10:17911"
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/SyncTaskStatus"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Response"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/SyncTaskRecord"

# ────────────────────────────────────────────────────────────────────────────
# Master: SyncRule CRUD + state transitions
# ────────────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: "`Authorization: Bearer <syncAdminToken>`"
    SyncTokenAuth:
      type: apiKey
      in: header
      name: X-Sync-Token
      description: Alternative to Bearer; same token value from `syncAdminToken` in master.json
  schemas:
...
    Response:
      type: object
      required: [code, msg]
      properties:
        code:
          type: integer
          example: 0
          description: "0 = success. Non-zero = error (check msg). HTTP status is always 200."
        msg:
          type: string
          example: "OK"
        data:
          nullable: true
          description: Payload on success; null on error.

    # ── SyncRule ───────────────────────────────────────────────────────────
...
    NodeLifecycleResponse:
      type: object
      properties:
        addr:
          type: string
        state:
          type: string
          enum: [active, draining]
        force:
          type: boolean
          description: Present on decommission responses only
        drained:
          type: integer
          description: Number of task cancel packets sent
        removed:
          type: boolean
          description: true when force=true and the node record was raft-deleted

    # ── SyncTask ───────────────────────────────────────────────────────────
    SyncTaskStatus:
      type: string
      enum: [queued, running, succeeded, failed, cancelled]
      description: |
        - `queued`: dispatched, not yet started on syncnode
        - `running`: actively executing on owning syncnode
        - `succeeded`: completed without error
        - `failed`: completed with error (see `error` field)
        - `cancelled`: cancelled by operator or fan-out parent
    SyncTaskProgress:
      type: object
      properties:
        filesTotal:
          type: integer
          format: int64
        filesDone:
          type: integer
          format: int64
        filesSkipped:
          type: integer
          format: int64
        filesFailed:
          type: integer
          format: int64
        bytesTotal:
          type: integer
          format: int64
        bytesDone:
          type: integer
          format: int64
        throughputMBps:
          type: number
          format: double
    SyncTaskRecord:
      type: object
      description: |
        Task ledger entry. Fan-out parent tasks have `owner=""` and
        `shardTotal > 0`. Their shard children carry `owner=<addr>` and
        `shardIdx`. `taskID` of shards is `<parentID>/<shardIdx>`.
      properties:
        taskID:
          type: string
          example: "backup-cfs-to-s3/1715798400000000000"
        ruleID:
          type: string
          description: ID of the originating rule; empty for tasks created via /syncNode/dispatch
        type:
          type: string
        status:
          $ref: "#/components/schemas/SyncTaskStatus"
        owner:
          type: string
          description: Syncnode addr owning this task; empty for fan-out parent tasks
          example: "10.89.6.10:17911"
        shardIdx:
          type: integer
          description: Zero-based shard index (fan-out children only)
        shardTotal:
          type: integer
          description: Total shard count (fan-out parent/children only)
        startedAt:
          type: string
          format: date-time
        doneAt:
          type: string
          format: date-time
          description: Zero / omitted until terminal
        error:
          type: string
          description: Non-empty only on status=failed
        progress:
          $ref: "#/components/schemas/SyncTaskProgress"

    # ── Dispatch ───────────────────────────────────────────────────────────
...